+++ test/chruby_auto_test.sh
@@ -1,5 +1,3 @@
-#!/bin/sh
-
 . ./share/chruby/auto.sh
 . ./test/helper.sh
 
@@ -11,19 +9,60 @@ setUp()
 	unset RUBY_VERSION_FILE
 }
 
+test_chruby_auto_setting_blank_PROMPT_COMMAND()
+{
+	if [[ -n "$BASH_VERSION" ]]; then
+		PROMPT_COMMAND=""
+		. ./share/chruby/auto.sh
+
+		assertEquals "has syntax error" \
+			     "chruby_auto" \
+			     "$PROMPT_COMMAND"
+	fi
+}
+
+test_chruby_auto_setting_PROMPT_COMMAND_with_semicolon()
+{
+	if [[ -n "$BASH_VERSION" ]]; then
+		PROMPT_COMMAND="update_terminal_cwd;"
+		. ./share/chruby/auto.sh
+
+		assertEquals "did not remove tailing ';' and whitespace" \
+			     "update_terminal_cwd; chruby_auto" \
+			     "$PROMPT_COMMAND"
+	fi
+}
+
+test_chruby_auto_setting_PROMPT_COMMAND_with_semicolon_and_whitespace()
+{
+	if [[ -n "$BASH_VERSION" ]]; then
+		PROMPT_COMMAND="update_terminal_cwd; "
+		. ./share/chruby/auto.sh
+	
+		assertEquals "did not remove tailing ';' and whitespace" \
+			     "update_terminal_cwd; chruby_auto" \
+			     "$PROMPT_COMMAND"
+	fi
+}
+
 test_chruby_auto_loaded_twice()
 {
+	RUBY_VERSION_FILE="dirty"
+	PROMPT_COMMAND="chruby_auto"
+
 	. ./share/chruby/auto.sh
 
 	if [[ -n "$ZSH_VERSION" ]]; then
 		assertNotEquals "should not add chruby_auto twice" \
 			        "$precmd_functions" \
 				"chruby_auto chruby_auto"
-	else
+	elif [[ -n "$BASH_VERSION" ]]; then
 		assertNotEquals "should not add chruby_auto twice" \
 			        "$PROMPT_COMMAND" \
 		                "chruby_auto; chruby_auto"
 	fi
+
+	assertNull "RUBY_VERSION_FILE was not unset" "$RUBY_VERSION_FILE"
 }
 
 test_chruby_auto_enter_project_dir()
@@ -31,7 +70,7 @@ test_chruby_auto_enter_project_dir()
 	cd "$PROJECT_DIR" && chruby_auto
 
 	assertEquals "did not switch Ruby when entering a versioned directory" \
-		     "$TEST_RUBY" "$RUBY"
+		     "$TEST_RUBY_ROOT" "$RUBY_ROOT"
 }
 
 test_chruby_auto_enter_subdir_directly()
@@ -39,7 +78,7 @@ test_chruby_auto_enter_subdir_directly()
 	cd "$PROJECT_DIR/sub_dir" && chruby_auto
 
 	assertEquals "did not switch Ruby when directly entering a sub-directory of a versioned directory" \
-		     "$TEST_RUBY" "$RUBY"
+		     "$TEST_RUBY_ROOT" "$RUBY_ROOT"
 }
 
 test_chruby_auto_enter_subdir()
@@ -48,7 +87,7 @@ test_chruby_auto_enter_subdir()
 	cd sub_dir        && chruby_auto
 
 	assertEquals "did not keep the current Ruby when entering a sub-dir" \
-		     "$TEST_RUBY" "$RUBY"
+		     "$TEST_RUBY_ROOT" "$RUBY_ROOT"
 }
 
 test_chruby_auto_enter_subdir_with_ruby_version()
@@ -57,7 +96,7 @@ test_chruby_auto_enter_subdir_with_ruby_version()
 	cd sub_versioned  && chruby_auto
 
 	assertNull "did not switch the Ruby when leaving a sub-versioned directory" \
-		   "$RUBY"
+		   "$RUBY_ROOT"
 }
 
 test_chruby_auto_overriding_ruby_version()
@@ -65,7 +104,7 @@ test_chruby_auto_overriding_ruby_version()
 	cd "$PROJECT_DIR" && chruby_auto
 	chruby system     && chruby_auto
 
-	assertNull "did not override the Ruby set in .ruby-version" "$RUBY"
+	assertNull "did not override the Ruby set in .ruby-version" "$RUBY_ROOT"
 }
 
 test_chruby_auto_leave_project_dir()
@@ -75,7 +114,7 @@ test_chruby_auto_leave_project_dir()
 	cd ../../..       && chruby_auto
 
 	assertNull "did not reset the Ruby when leaving a versioned directory" \
-		   "$RUBY"
+		   "$RUBY_ROOT"
 }
 
 test_chruby_auto_invalid_ruby_version()
@@ -84,7 +123,7 @@ test_chruby_auto_invalid_ruby_version()
 	cd bad            && chruby_auto 2>/dev/null
 
 	assertEquals "did not keep the current Ruby when loading an unknown version" \
-		     "$TEST_RUBY" "$RUBY"
+		     "$TEST_RUBY_ROOT" "$RUBY_ROOT"
 }
 
 SHUNIT_PARENT=$0 . $SHUNIT2